Add files via upload
While running the script by visualcurrent, I have noticed that the UID was not removed from the files and folders automatically. Arozx has fixed this issue in his fork, however it was not compatible with macOS as the paths were preconfigured. This is why I have removed the preconfiguration, so the user can use the path of choice.
Also, the headings were doubled as the first line also contained the heading with a "#", which I did not like. This is why I have excluded the first line in this script.